+++ src/function.ts
@@ -628,12 +628,3 @@ export function pipe(
  * @since 2.7.0
  */
 export const hole: <T>() => T = absurd as any
-
-/**
- * @internal
- */
-export const bind_ = <A, N extends string, B>(
-  a: A,
-  name: Exclude<N, keyof A>,
-  b: B
-): { [K in keyof A | N]: K extends keyof A ? A[K] : B } => Object.assign({}, a, { [name]: b }) as any